Fixed first time input errors
Free Flight would throw exceptions when it tried to get inputs for flare or dive controls.
This was a problem, because these controls wouldn't exist yet, and the user was given
errors the first time they imported Free Flight. 
 
These are now disabled by default, and enabled when the user decides to set them up.