Major Refactor: Decoupled main systems
The main Free Flight class was becoming bloaded and needed to be split
up. Since version 0.4, the prototyping stages have ended and the code is
being refactored into a production environment. The new major systems
include: Mode, and Inputs classes. Modes are responsible for handling
separate forms of movement, such as flying and ground. The inputs classes
further provide player input support to the Modes.