Refactored and overhauled the mechanic execution system
Specifically, how and which mechanics are executed over another? The process
is now automated and depends on which mechanic is higher in the list. The 
mechanics themselves are still hard-coded into the list, and that won't change
until an editor is written for the list to be edited by the user. 

This process is far from finished, but it's preferable to the last prototype which 
specifically hardcoded which mechanics could run over another. Eventually, users
should be able to mix and match different flight mechanics without looking at code.