Added failed dynamic mechanic editor
This adds the beginning of an editor designed to handle custom
mechanics in the editor. It's a "failed attempt" because the editor
can't handle serializing polymorphic derived classes, so there isn't
really any point to adding this editor functionality. It will be committed
anyway because it *will* work if Unity adds support for serializing 
polymorphic classes. 

What does this mean? Mechanics are dynamically addable, but 
are not fully editable via the inspector. Only types belonging to 
the base class Mechanic can be changed.