Improved the derivation of Lift Coefficient
Added new linear backup equation for calculating outlire angles of 
attack.

The last model is an inverted quadratic, which started to fail
at very high angles of attack. Broadening the slope of the 
equation caused the lift coefficient to be less dynamic. The 
linear backup equation keeps the dynamic behaviour of the
quadratic model and allows for broader angles of attack to happen.